reduced memory consumption and split vis / simul code
