+++ src/models/defaults.py
@@ -111,6 +111,8 @@ default_plot_ylim_cut_top = None
 default_plot_xlim_cut_left = None
 default_plot_ylim_cut_bottom = None
 
+default_enable_visualization = False
+
 defaults = {
     INITIAL_CONDITIONS: default_initial_conditions,
     STOP_SIMULATION_THRESHOLD: default_stop_simulation_threshold,
@@ -149,7 +151,7 @@ defaults = {
     PLOT_YLIM_CUT_TOP: default_plot_ylim_cut_top,
     PLOT_YLIM_BOTTOM: default_plot_ylim_bottom,
     PLOT_YLIM_TOP: default_plot_ylim_top,
-
+    ENABLE_VISUALIZATION: default_enable_visualization
 }
 
 default_age_induced_fatality_rates = [(0, 20, 0.002), (20, 40, 0.002), (40, 50, 0.004), (50, 60, 0.013),
